Refactor locations of code in dx to bundle stuff at a sane place.
