+++ src/plone/app/multilingual/dx/languageindependent.py
@@ -4,15 +4,47 @@ from plone.app.multilingual.interfaces import ILanguage
 from plone.app.multilingual.interfaces import ILanguageIndependentFieldsManager
 from plone.app.multilingual.interfaces import ITranslationManager
 from plone.dexterity.utils import iterSchemata
+from z3c.form.interfaces import DISPLAY_MODE
+from z3c.form.validator import StrictSimpleFieldValidator
 from z3c.relationfield import RelationValue
 from z3c.relationfield.interfaces import IRelationList
 from z3c.relationfield.interfaces import IRelationValue
 from zope.app.intid.interfaces import IIntIds
+from zope.component import getMultiAdapter
 from zope.component import getUtility
 from zope.component import queryAdapter
 from zope.interface import implementer
+from zope.pagetemplate.interfaces import IPageTemplate
 
 _marker = object()
+
+
+class LanguageIndependentFieldValidator(StrictSimpleFieldValidator):
+    """Override validator so we can ignore language independent fields,
+       these will be automatically filled later on by subscriber.createdEvent
+    """
+    def validate(self, value, force=False):
+        # always pass
+        pass
+
+
+class LanguageIndependentFieldInputTemplate(object):
+    """Override input template for language independent fields with
+       display widget, because values will be automatically filled
+       by later on by subscriber.createdEvent.
+    """
+    def __init__(self, context, request, view, field, widget):
+        self.context = context
+        self.request = request
+        self.view = view
+        self.field = field
+        self.widget = widget
+
+    def __call__(self, widget):
+        template = getMultiAdapter(
+            (self.context, self.request, self.view, self.field, self.widget,),
+            IPageTemplate, name=DISPLAY_MODE)
+        return template(widget)
 
 
 @implementer(ILanguageIndependentFieldsManager)
